{"schema":"decodex/commit/1","summary":"Modularize macOS cursor runtime","authority":"manual"}
