+++ packages/rsnap-overlay/src/overlay/rendering.rs
@@ -32,9 +32,9 @@ use winit::window::Window;
 use self::hud_rendering::LiveLoupeTexture;
 use self::hud_surface::{HudBg, HudBlurUniformRaw};
 #[cfg(target_os = "macos")]
-use crate::overlay::MacOSOverlayCursorRectSupport;
-#[cfg(target_os = "macos")]
 use crate::overlay::macos_configure_hud_window;
+#[cfg(target_os = "macos")]
+use crate::overlay::macos_cursor_runtime::MacOSOverlayCursorRectSupport;
 use crate::overlay::{
 	self, AcquiredSurfaceFrame, Adapter, AddressMode, Arc, BindGroupLayout, BindingResource,
 	BindingType, BlendState, Buffer, BufferBindingType, BufferSize, BufferUsages, ClippedPrimitive,
